Batch writes to DB
The writes to the database are now batched. There is a max batch size,
when it is reached, the elements to insert are written to the  DB. There
is also a max time intervall, after which a batch is written, even if
the number of elements to insert  is lower than the max batch size.

In addition, this commit swichtes from the Apache Camel JDBC  to
the Apache Camel SQL component for database interaction.